requirements.txt
torch
torchvision
segmentation-models-pytorch
matplotlib
numpy
Pillow
streamlit
